+++ NEWS.md
@@ -1,5 +1,9 @@
 # tensorflow (development version)
 
+- `as_tensor()` gains a `shape` argument, can be used to fill or reshape the supplied object.
+  Scalars can be recycled to a tensor of arbitrary `shape`, otherwise 
+  supplied objects are reshaped using row-major (C-style) semantics.
+ 
 - `install_tensorflow()` now provides experimental support for Arm Macs,
   with the following restrictions:
     - "conda" is the only supported installation method.